feat: add infinite scroll to RecyclerView
