+++ app/src/main/AndroidManifest.xml
@@ -16,20 +16,20 @@
         android:usesCleartextTraffic="true"
         tools:ignore="Instantiatable">
 
-        <activity android:name=".feature.navigation.home.NewsActivity"></activity>
-        <activity android:name=".feature.law.DetailResultActivity" />
-        <activity android:name=".feature.law.search_result.SearchResultActivity" />
-        <activity android:name=".feature.navigation.user.UserInfoEditActivity" />
-        <activity android:name=".feature.navigation.MainActivity">
+        <activity android:name=".ui.navigation.home.NewsActivity"></activity>
+        <activity android:name=".ui.law.DetailResultActivity" />
+        <activity android:name=".ui.law.search_result.SearchResultActivity" />
+        <activity android:name=".ui.navigation.user.UserInfoEditActivity" />
+        <activity android:name=".ui.navigation.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".feature.SplashActivity" />
-        <activity android:name=".feature.signup.SignUpActivity" />
-        <activity android:name=".feature.signin.SignInActivity" />
+        <activity android:name=".ui.SplashActivity" />
+        <activity android:name=".ui.signup.SignUpActivity" />
+        <activity android:name=".ui.signin.SignInActivity" />
     </application>
 
 </manifest>
