+++ app/src/main/java/com/example/judgement/ui/navigation/scrap/ScrapManager.kt
@@ -1,4 +1,4 @@
-package com.example.judgement.feature.navigation.scrap
+package com.example.judgement.ui.navigation.scrap
 
 data class ScrapManager(
     val scrapRvAdapters: MutableMap<String, ScrapRvAdapter> = mutableMapOf(), // 어댑터
