+++ app/src/main/java/com/example/judgement/ui/navigation/scrap/ScrapPagerAdapter.kt
@@ -1,12 +1,10 @@
-package com.example.judgement.feature.navigation.scrap
+package com.example.judgement.ui.navigation.scrap
 
 import android.content.Context
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
-import android.widget.Toast
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
