+++ app/src/main/res/layout/activity_news.xml
@@ -7,13 +7,13 @@
     <data>
         <variable
             name="activity_news"
-            type="com.example.judgement.feature.navigation.home.NewsActivity" />
+            type="com.example.judgement.ui.navigation.home.NewsActivity" />
     </data>
 
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".feature.navigation.home.NewsActivity">
+        tools:context=".ui.navigation.home.NewsActivity">
 
         <WebView
             android:id="@+id/web_view_news"
