+++ app/src/main/res/layout/activity_sign_up.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="activity"
-            type="com.example.judgement.feature.signup.SignUpActivity" />
+            type="com.example.judgement.ui.signup.SignUpActivity" />
 
     </data>
